# docs/swagger.yaml
definitions:
  global.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      traceId:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: 25076778@qq.com
    name: API Support
    url: http://www.swagger.io/support
  description: Gin-base API 文档
  termsOfService: http://swagger.io/terms/
  title: Gin-base Swagger API
  version: "2.0"
paths:
...
  /api/v1/role:
    get:
      description: 角色列表
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 分页大小
        in: query
        name: pageSize
        required: true
        type: string
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 列表
      tags:
      - 角色管理
    post:
      description: 角色创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 创建
      tags:
      - 角色管理
  /api/v1/role/{id}:
    get:
      description: 角色详情
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 详情
      tags:
      - 角色管理
    put:
      description: 角色更新
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 更新
      tags:
      - 角色管理
...
  /api/v1/user:
    get:
      description: 用户列表
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 分页大小
        in: query
        name: pageSize
        required: true
        type: string
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 列表
      tags:
      - 用户管理
    post:
      description: 用户创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 创建
      tags:
      - 用户管理
  /api/v1/user/{id}:
    get:
      description: 用户详情
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 详情
      tags:
      - 用户管理
    put:
      description: 用户更新
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 更新
      tags:
      - 用户管理
  /v1/role/{id}:
    delete:
      description: 角色删除
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 删除
      tags:
      - 角色管理
  /v1/user/{id}:
    delete:
      description: 用户删除
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 删除
      tags:
      - 用户管理
swagger: "2.0"